updating install script

+++ install.sh
@@ -4,7 +4,7 @@ function install() {
     git clone https://github.com/jacbart/dotfiles.git $HOME/.dotfiles
 
     if type zsh &> /dev/null; then
-        rm -rf  $HOME/.zshrc
+        rm -f  $HOME/.zshrc
         ln -s $HOME/.dotfiles/shell/zshrc $HOME/.zshrc
         touch $HOME/.zshrc.local
 
@@ -12,12 +12,13 @@ function install() {
         [[ -d $HOME/bin ]] || mkdir $HOME/bin
         
         if type nvim &> /dev/null; then
-            rm -rf  $HOME/.config/nvim
+            [[ -d $HOME/.config/nvim ]] && rm -rf  $HOME/.config/nvim
             mkdir -p $HOME/.config/nvim
             ln -s $HOME/.dotfiles/config/init.vim $HOME/.config/nvim/init.vim
         fi
 
         if type kitty &> /dev/null; then
+            [[ -d $HOME/.config/kitty ]] && rm -rf  $HOME/.config/kitty
             mkdir -p $HOME/.config/kitty
             ln -s $HOME/.dotfiles/config/kitty.conf $HOME/.config/kitty/kitty.conf
         fi
@@ -40,7 +41,7 @@ function yes_or_no {
 
 if [[ -d $HOME/.dotfiles ]]; then
     message="~/.dotfiles already exists, would you like to remove and install again?"
-    yes_or_no "$message" && install
+    yes_or_no "$message" && rm -rf "$HOME/.dotfiles" && install
 else
     install
 fi
